Issue #70 (moved CompileMain to compiler project)
