+++ webapp/src/main/java/info/teksol/mindcode/webapp/HomeController.java
@@ -20,7 +20,7 @@ import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
-import static info.teksol.mindcode.webapp.CompilerFacade.compile;
+import static info.teksol.mindcode.mindustry.CompilerFacade.compile;
 
 @Controller
 @RequestMapping(value = "/")
